Use CocoaPods resource bundle for StreamChat
